Session manages output and only interacts dynamically replace output
Session object now includes mehtods to manage where output is outputted as well as whether the output should replace or be appended to existing output and whether calling the setOutput() method again (without resetting the output settings first) is allowed to overwrite the current configuration.
Session object also is in charge of outputting content.
The sendMsg() function in the session object now accepts a variable number of arguments, the second of which is a custom message id.
Interacts dynamically replace their own output, but all other output is appended.