More embedding options.
Now supports showFileUploads = [javascript boolean] and showComputationID = [javascript boolean] to allow for customized embedding.
This allows someone to show only the codemirror box (with "eval" and "sagemode" buttons) and the output div for results when they embed it.

Also implements move() and restore() functions that can be called before the submission of an encapsulating form, so as to prevent extra requests to the server (e.g. file uploads).