Interact output bugfix
Previously, if a complex interact was updated multiple times quicker than the server can reply, the output would be printed multiple times. This fixes the issue by having the session object track the last sent request to the server and if the messages it gets back are in response to a different request, output is not formatted (but the messages are logged in the messages div).