Fix failing tests
test_add_computer_success was failing because we don't use a direct
socket to communicate with the untrusted side anymore, so it wasn't
stored in the expected location.

test_interrupt_kernel was failing because the message being sent to the
untrusted side had a spelling error in the message type.