Only use SockJS for connections for now.  There are some issues and race conditions with using straight websockets.
Before this commit, if we have a very short computation (like 1+1), there is some sort of 
race condition where the iopub handler does not get established before 
the kernel is closed down.  This only manifests itself on a remote server, since presumably
if you are running on a local server, the connection is established too quickly.

Also, there are some bugs in, for example, Firefox and other issues that we don't want to have
to work around, that sockjs already worked around.  See the SockJS websocket transport.