Provide an option (sagecell.loadMathJax) to control whether the sage cell loads mathjax.
Set this to false *before* making any sage cells to prevent the sagecell from loading mathjax.  This is very useful if the page already loads mathjax.  The user takes responsibility for loading mathjax with a compatible configuration to the sage cell's.

We do this because mathjax does *not* like to be loaded twice with different configurations.

Also, there is a check to see if mathjax is loaded already, and if it is, don't load it again.