Fix interacts and new files with the new IPython.
IPython commit d7db8e492 changed the user_expressions return format to a rich data display, so we had to update that.

For new files, we weren't including the payload dictionary in interact updates, and the message handler wasn't acting on new files.  Also, now we send an update message with every control update, whether or not the interact is actually run, so we now make sure we actually have new files before pulsating the file list.