Comment out old files code
This code is in here because SageCell version 1 let a user upload files to act upon.  When we rewrote the cell server, we decided uploading files was a hardly-used feature, so we put it on the back burner and never got around to implementing it.  It is potentially useful, though, and this code works around some browser problems, so I'm leaving the code in, but commented out.