Fix errors for rotation
As best as I can tell, the rolling calculations would make it so that our particular dom element would consistently roll around a particular axis.  This may have been an effect of our doom element being the width of the browser, but not very high.

I also modified the trackball control to trigger when there are changes, and rewrote the animation code to only animate when there are possible control changes.