Fix z-fighting issue
Sage checks to see if an object is enclosed, and if it is, Sage doesn't bother drawing the back faces to polygons.  If an object is not enclosed, Sage explicitly draws both the back and front face of the polygons.

A better solution, I think, would be to have Sage not draw the back faces to polygons, or at least have an option to turn it off.