Fix rotating lights to scale distances properly.
If we leave the rotating lights alone, then when the distance between the camera and the target scales, the lights hit different points on the object.  Instead, we now scale the distance between the light and the camera to be proportional to the distance between the camera and the object.  This ensures that the light is pointed at the same angle with respect to the object regardless of distance to the object.