Add a retry option to the sagecell service checking
It seems that every now and then, this service request returns a success message, but there is no 'stout' key in the reply dictionary, which means that the service is restarted every couple of hours.  We were able to reproduce this sort of failure on personal machines by throwing hundreds of similar service requests at a server, and had maybe one failure every 100 requests or so.  Diving deeper, it appeared that for some reason, the zmq message of the computation was not getting sent to the kernel, even though it claimed that it was sent.  Upgrading to zmq 4.x didn't seem to help.  While we investigate further, we'll put in this stopgap measure that just retries the request after a short delay.