Create CellSessionID to help load balancer.
This ID is passed as a parameter in kernel and sockjs connection requests.
In the same table of HA-Proxy we store Jupyter kernel IDs and for now
rewrite URLs of file requests to include the same CellSessionID parameter
as their kernel ID present in the file path.

This change should greatly improve load balancing compared to IP sticking
and decrease likelyhood for active users to be stuck with an old container
that gets destroyed during an update.