Support nontrivial GAP input
I have not yet tested this change; I'm submitting it just to ask if something this straightforward could solve the issue in [this post](https://groups.google.com/d/msg/sage-devel/VEiahzqU2qs/sT7uf9efCwAJ) (which Andrey has been kind enough to ask on my behalf).  I tried to test this with a docker container of SageMathCell but the one I found was not functioning.  If you'd like to test this, or teach me how to test it, that could be good.  Or this can be the start of a discussion.